update: テーマ切り替え機能改修とTypeScriptエラー改修
