+++ frontend/app/src/App.tsx
@@ -1,9 +1,9 @@
 import { type FC } from 'react';
 import './App.css';
+import { ThemeProvider } from 'components/Theme/themeProvider';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { ThemeSwitchButton } from 'components/Theme/ThemeSwitchButton';
 import { Sidebar } from './components/Sidebar/Sidebar';
-import { ThemeSwitchButton } from './components/Theme2/ThemeSwitchButton';
-// import { ThemeProvider } from './components/Theme2/themeProvider';
 
 import Overview from './pages/Overview';
 import Settings from './pages/Settings';
@@ -11,17 +11,17 @@ import Settings from './pages/Settings';
 const App: FC = () => {
   return (
     <>
-      {/* <ThemeProvider> */}
-      <BrowserRouter>
-        <Sidebar />
+      <ThemeProvider>
+        <BrowserRouter>
+          <Sidebar />
 
-        <ThemeSwitchButton />
-        <Routes>
-          <Route path="/overview" element={<Overview />} />
-          <Route path="/settings" element={<Settings />} />
-        </Routes>
-      </BrowserRouter>
-      {/* </ThemeProvider> */}
+          <ThemeSwitchButton />
+          <Routes>
+            <Route path="/overview" element={<Overview />} />
+            <Route path="/settings" element={<Settings />} />
+          </Routes>
+        </BrowserRouter>
+      </ThemeProvider>
     </>
   );
 };
