--- frontend/app/src/components/Theme/styles/index.ts
@@ -1,2 +1,0 @@
-export * from './globalStyle';
-export * from './theme';
