+++ frontend/app/src/pages/Settings.tsx
@@ -1,7 +1,22 @@
 import { type FC } from 'react';
+// import { useTheme } from '../components/Theme/themeContext';
 
 const Settings: FC = () => {
-  return <div className="overview">Settings!</div>;
+  // const { colorMode, setColorMode } = useTheme();
+
+  // return (
+  //   <>
+  //     <div className="overview">Settings!</div>
+  //     <p>current color mode:{colorMode}</p>
+  //     <button onClick={setColorMode}>toggle color mode</button>
+  //   </>
+  // );
+
+  return (
+    <>
+      <div className="overview">Settings!</div>;
+    </>
+  );
 };
 
 export default Settings;
